+++ tests/app/models/filters.py
@@ -1,3 +1,4 @@
+from django.core.validators import MaxValueValidator
 from django.db import models
 from django.utils.timezone import now
 
@@ -24,10 +25,14 @@ class Message(models.Model):
         ("video", "Video"),
     )
 
-    sender = models.ForeignKey(User, on_delete=models.CASCADE, related_name="sent_messages")
+    sender = models.ForeignKey(User, on_delete=models.CASCADE, null=True, related_name="sent_messages")
     group = models.ForeignKey(Group, on_delete=models.CASCADE, related_name="messages")
     type = models.CharField(max_length=10, choices=MESSAGE_TYPES, default="text")
     text = models.TextField()
+    rating = models.PositiveIntegerField(default=0, validators=[
+        MaxValueValidator(5)
+    ])
+    edited = models.BooleanField(default=False, null=True)
     created_at = models.DateTimeField(default=now)
 
     def __str__(self):